Item manager #99 and #88

+++ src/cam.cpp
@@ -920,8 +920,8 @@ void Cam::resume() {
         //        std::cout << "Resumed" << std::endl;
     }
 }
-void Cam::seek(float time) {
-    std::chrono::nanoseconds nano(static_cast<ulong>(1e9f*time));
+void Cam::seek(double time) {
+    std::chrono::nanoseconds nano(static_cast<ulong>(1e9*time));
     static_cast<rs2::playback>(dev).seek(nano);
 }
